refactor: extract shared compaction module, add to CLI agent (4c)
Move build_tool_stub and compact_tool_messages from api_server.py into
services/compaction.py. CLI agent now compacts tool messages after each
round to prevent unbounded token growth. Enhanced build_tool_stub to
handle content and date fields from standardized responses.
